Bug 468994 - fixing warning about formal parameter 2 different from declaration int16 != int, r=joe
UltraBlame original commit: 6956e3df30f564c75070ec580a7fb047014dd1a2

+++ jpeg/jcmarker.c
@@ -433,7 +433,7 @@ emit_adobe_app14 (j_compress_ptr cinfo)
 
 
 METHODDEF(void)
-write_marker_header (j_compress_ptr cinfo, int16 marker, unsigned int datalen)
+write_marker_header (j_compress_ptr cinfo, int marker, unsigned int datalen)
 
 {
   if (datalen > (unsigned int) 65533)		
@@ -445,10 +445,10 @@ write_marker_header (j_compress_ptr cinfo, int16 marker, unsigned int datalen)
 }
 
 METHODDEF(void)
-write_marker_byte (j_compress_ptr cinfo, int16 val)
+write_marker_byte (j_compress_ptr cinfo, int val)
 
 {
-  emit_byte(cinfo, val);
+  emit_byte(cinfo, (int16) val);
 }
